actualizada el get de sale
