--- api/src/controllers/Sale/getSales.js
@@ -1,8 +1,0 @@
-const { Sale } = require("../../db");
-
-const getSaleById = async (id) => {
-    const sale = await Sale.findByPk(id);
-    return sale;
-};
-
-module.exports = getSaleById;
