+++ api/src/handlers/saleHandler.js
@@ -1,6 +1,5 @@
-const getSaleById = require("../controllers/Sale/getSales");
+const getSaleById = require("../controllers/Sale/getSaleById");
 const getMySales = require("../controllers/Sale/getMySales");
-const {Order, Event, User} = require ('../db')
 
 const getMySalesHandler = async (req, res) => {
     try {
@@ -13,37 +12,13 @@ const getMySalesHandler = async (req, res) => {
 };
 
 const getSaleByIdHandler = async (req, res) => {
-    const { id } = req.params;  
-
-    
     try {
+        const { id } = req.params;  
         const saleById = await getSaleById(id);
-        if(!saleById){
-            throw Error("No se encuentra venta")
-        }else{
-            const order = await Order.findByPk(saleById.idOrder)
-            const user = await User.findByPk(order.idBuyer);
-            const event = await Event.findByPk(order.idEvent);
-
-            res.status(200).json( {
-                name: user.name,
-                email: user.email,
-                eventName: event.name,
-                eventImage: event.image,
-                price: order.price,
-                quantity: order.quantity,
-                currency: event.currency,
-                date: event.date,
-                hour: event.hour,
-                address: event.address,
-                country: event.country,
-                ...saleById
-             });
+        res.status(200).json(saleById);
         
-        }
-         
    } catch (error) {
-        res.status(404).send(error.message)
+        res.status(404).json({error: error.message});
     }
 };
